# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI-Powered Travel Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /flights/search:
    get:
      operationId: searchFlights
      tags: [flights]
      summary: Search for flights
      parameters:
        - name: origin
          in: query
          required: true
          schema:
            type: string
        - name: destination
          in: query
          required: true
          schema:
            type: string
        - name: departureDate
          in: query
          required: true
          schema:
            type: string
        - name: returnDate
          in: query
          schema:
            type: string
        - name: passengers
          in: query
          schema:
            type: integer
        - name: cabinClass
          in: query
          schema:
            type: string
            enum: [economy, premium_economy, business, first]
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: stops
          in: query
          schema:
            type: string
            enum: [any, nonstop, one_stop]
      responses:
        "200":
          description: Flight search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlightSearchResults"
  /flights/price-calendar:
    get:
      operationId: getFlightPriceCalendar
      tags: [flights]
      summary: Get price calendar for a route
      parameters:
        - name: origin
          in: query
          required: true
          schema:
            type: string
        - name: destination
          in: query
          required: true
          schema:
            type: string
        - name: month
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Price calendar
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PriceCalendarEntry"

  # ==================== HOTEL SEARCH ====================
  /hotels/search:
    get:
      operationId: searchHotels
      tags: [hotels]
      summary: Search for hotels
      parameters:
        - name: destination
          in: query
          required: true
          schema:
            type: string
        - name: checkIn
          in: query
          required: true
          schema:
            type: string
        - name: checkOut
          in: query
          required: true
          schema:
            type: string
        - name: guests
          in: query
          schema:
            type: integer
        - name: minStars
          in: query
          schema:
            type: integer
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: amenities
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Hotel search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Hotel"
  /hotels/{id}:
    get:
      operationId: getHotel
      tags: [hotels]
      summary: Get hotel details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Hotel detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HotelDetail"

  # ==================== ACTIVITIES ====================
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expenses
      parameters:
        - name: tripId
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
        - name: reportId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseBody"
      responses:
        "201":
          description: Created expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExpenseBody"
      responses:
        "200":
          description: Updated expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /expenses/reports:
    get:
      operationId: listExpenseReports
      tags: [expenses]
      summary: List expense reports
      responses:
        "200":
          description: List of expense reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExpenseReport"
    post:
      operationId: createExpenseReport
      tags: [expenses]
      summary: Create an expense report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseReportBody"
      responses:
        "201":
          description: Created report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseReport"
  /expenses/summary:
    get:
      operationId: getExpenseSummary
      tags: [expenses]
      summary: Get expense summary by category
      parameters:
        - name: tripId
          in: query
          schema:
            type: integer
        - name: period
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Expense summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseSummary"

  # ==================== PROFILE ====================
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /profile/documents:
    get:
      operationId: listDocuments
      tags: [profile]
      summary: List travel documents
      responses:
        "200":
          description: Travel documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TravelDocument"
    post:
      operationId: createDocument
      tags: [profile]
      summary: Add a travel document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentBody"
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelDocument"
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    FlightSearchResults:
      type: object
      properties:
        outbound:
          type: array
          items:
            $ref: "#/components/schemas/Flight"
        return:
          type: array
          items:
            $ref: "#/components/schemas/Flight"
        meta:
          $ref: "#/components/schemas/SearchMeta"
      required:
        - outbound
        - return
        - meta
    Flight:
      type: object
      properties:
        id:
          type: string
        airline:
          type: string
        airlineCode:
          type: string
        airlineLogo:
          type: ["string", "null"]
        flightNumber:
          type: string
        origin:
          type: string
        destination:
          type: string
        departure:
          type: string
        arrival:
          type: string
        duration:
          type: integer
        stops:
          type: integer
        price:
          type: number
        currency:
          type: string
        cabinClass:
          type: string
        baggageIncluded:
          type: boolean
        refundable:
          type: boolean
        dealScore:
          type: number
        carbonKg:
          type: number
        seatsLeft:
          type: ["integer", "null"]
      required:
        - id
        - airline
        - airlineCode
        - airlineLogo
        - flightNumber
        - origin
        - destination
        - departure
        - arrival
        - duration
        - stops
        - price
        - currency
        - cabinClass
        - baggageIncluded
        - refundable
        - dealScore
        - carbonKg
        - seatsLeft
    PriceCalendarEntry:
      type: object
      properties:
        date:
          type: string
        price:
          type: number
        available:
          type: boolean
        priceLevel:
          type: string
          enum: [low, medium, high]
      required:
        - date
        - price
        - available
        - priceLevel
    SearchMeta:
      type: object
      properties:
        total:
          type: integer
        cheapest:
          type: number
        fastest:
          type: integer
        bestDeal:
          type: string
      required:
        - total
        - cheapest
        - fastest
        - bestDeal
    Hotel:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        destination:
          type: string
        address:
          type: string
        stars:
          type: integer
        reviewScore:
          type: number
        reviewCount:
          type: integer
        pricePerNight:
          type: number
        currency:
          type: string
        imageUrl:
          type: ["string", "null"]
        amenities:
          type: array
          items:
            type: string
        freeCancellation:
          type: boolean
        aiSummary:
          type: ["string", "null"]
        dealScore:
          type: number
        distanceFromCenter:
          type: number
        neighborhood:
          type: ["string", "null"]
      required:
        - id
        - name
        - destination
        - address
        - stars
        - reviewScore
        - reviewCount
        - pricePerNight
        - currency
        - imageUrl
        - amenities
        - freeCancellation
        - aiSummary
        - dealScore
        - distanceFromCenter
        - neighborhood
    HotelDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        destination:
          type: string
        address:
          type: string
        stars:
          type: integer
        reviewScore:
          type: number
        reviewCount:
          type: integer
        pricePerNight:
          type: number
        currency:
          type: string
        imageUrl:
          type: ["string", "null"]
        amenities:
          type: array
          items:
            type: string
        freeCancellation:
          type: boolean
        aiSummary:
          type: ["string", "null"]
        dealScore:
          type: number
        distanceFromCenter:
          type: number
        neighborhood:
          type: ["string", "null"]
        rooms:
          type: array
          items:
            $ref: "#/components/schemas/HotelRoom"
        pros:
          type: array
          items:
            type: string
        cons:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - destination
        - address
        - stars
        - reviewScore
        - reviewCount
        - pricePerNight
        - currency
        - imageUrl
        - amenities
        - freeCancellation
        - aiSummary
        - dealScore
        - distanceFromCenter
        - neighborhood
        - rooms
        - pros
        - cons
    HotelRoom:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        pricePerNight:
          type: number
        maxGuests:
          type: integer
        bedType:
          type: string
        available:
          type: boolean
      required:
        - id
        - name
        - pricePerNight
        - maxGuests
        - bedType
        - available
...
    Expense:
      type: object
      properties:
        id:
          type: integer
        tripId:
          type: ["integer", "null"]
        reportId:
          type: ["integer", "null"]
        category:
          type: string
          enum: [accommodation, flights, transport, meals, activities, entertainment, telecoms, office, other]
        description:
          type: string
        amount:
          type: number
        currency:
          type: string
        amountUsd:
          type: number
        date:
          type: string
        merchant:
          type: ["string", "null"]
        receiptUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected, reimbursed]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - tripId
        - reportId
        - category
        - description
        - amount
        - currency
        - amountUsd
        - date
        - merchant
        - receiptUrl
        - status
        - notes
        - createdAt
    CreateExpenseBody:
      type: object
      properties:
        tripId:
          type: ["integer", "null"]
        reportId:
          type: ["integer", "null"]
        category:
          type: string
        description:
          type: string
        amount:
          type: number
        currency:
          type: string
        date:
          type: string
        merchant:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - category
        - description
        - amount
        - currency
        - date
    UpdateExpenseBody:
      type: object
      properties:
        category:
          type: string
        description:
          type: string
        amount:
          type: number
        currency:
          type: string
        date:
          type: string
        merchant:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
    ExpenseReport:
      type: object
      properties:
        id:
          type: integer
        tripId:
          type: ["integer", "null"]
        title:
          type: string
        status:
          type: string
          enum: [draft, submitted, under_review, approved, paid, rejected]
        totalAmount:
          type: number
        currency:
          type: string
        expenseCount:
          type: integer
        submittedAt:
          type: ["string", "null"]
        approvedAt:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - tripId
        - title
        - status
        - totalAmount
        - currency
        - expenseCount
        - submittedAt
        - approvedAt
        - paidAt
        - createdAt
    CreateExpenseReportBody:
      type: object
      properties:
        tripId:
          type: ["integer", "null"]
        title:
          type: string
        expenseIds:
          type: array
          items:
            type: integer
      required:
        - title
    ExpenseSummary:
      type: object
      properties:
        totalAmount:
          type: number
        currency:
          type: string
        byCategory:
          type: array
          items:
            $ref: "#/components/schemas/ExpenseCategoryBreakdown"
        byDay:
          type: array
          items:
            $ref: "#/components/schemas/ExpenseDayBreakdown"
        budgetUsed:
          type: ["number", "null"]
        budgetRemaining:
          type: ["number", "null"]
      required:
        - totalAmount
        - currency
        - byCategory
        - byDay
        - budgetUsed
        - budgetRemaining
    ExpenseCategoryBreakdown:
      type: object
      properties:
        category:
          type: string
        amount:
          type: number
        percentage:
          type: number
        count:
          type: integer
      required:
        - category
        - amount
        - percentage
        - count
    ExpenseDayBreakdown:
      type: object
      properties:
        date:
          type: string
        amount:
          type: number
      required:
        - date
        - amount
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        displayName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        nationality:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        phoneNumber:
          type: ["string", "null"]
        preferredCurrency:
          type: string
        preferredLanguage:
          type: string
        seatPreference:
          type: ["string", "null"]
        mealPreference:
          type: ["string", "null"]
        notificationsEnabled:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - email
        - firstName
        - lastName
        - displayName
        - avatarUrl
        - nationality
        - dateOfBirth
        - phoneNumber
        - preferredCurrency
        - preferredLanguage
        - seatPreference
        - mealPreference
        - notificationsEnabled
        - createdAt
    UpdateProfileBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        displayName:
          type: ["string", "null"]
        nationality:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        phoneNumber:
          type: ["string", "null"]
        preferredCurrency:
          type: string
        preferredLanguage:
          type: string
        seatPreference:
          type: ["string", "null"]
        mealPreference:
          type: ["string", "null"]
        notificationsEnabled:
          type: boolean
    TravelDocument:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [passport, national_id, visa, vaccination, drivers_license, other]
        name:
          type: string
        number:
          type: ["string", "null"]
        issuingCountry:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
        daysUntilExpiry:
          type: ["integer", "null"]
        status:
          type: string
          enum: [valid, expiring_soon, expired]
      required:
        - id
        - type
        - name
        - number
        - issuingCountry
        - expiryDate
        - daysUntilExpiry
        - status
    CreateDocumentBody:
      type: object
      properties:
        type:
          type: string
        name:
          type: string
        number:
          type: ["string", "null"]
        issuingCountry:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
      required:
        - type
        - name
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [booking_confirmation, price_drop, check_in_reminder, flight_status, visa_expiry, approval_required, safety_alert, points_expiry, general]
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        priority:
          type: string
          enum: [low, medium, high, critical]
        actionUrl:
          type: ["string", "null"]
        actionLabel:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - type
        - title
        - message
        - isRead
        - priority
        - actionUrl
        - actionLabel
        - createdAt